// src/shared/constant/error.enum.ts
// Note: enums ordered by its closes http error, from 2xx to 5xx
enum ErrorType {
  INVALID_INPUT = 'INVALID_INPUT',
  DATA_NOT_FOUND = 'DATA_NOT_FOUND',
  CONFLICT = 'CONFLICT',
  UNPROCESSABLE_ENTITY = 'UNPROCESSABLE_ENTITY',
  FAILED_DEPENDENCY = 'FAILED_DEPENDENCY',
  TOO_MANY_REQUESTS = 'TOO_MANY_REQUESTS',
  SERVICE_UNAVAILABLE = 'SERVICE_UNAVAILABLE',
  UNHANDLED_ERROR = 'UNHANDLED_ERROR',
  PAYMENT_REQUIRED_ERROR = 'PAYMENT_REQUIRED_ERROR',
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN'
}

export default ErrorType